rituraj added

+++ backend/controllers/dish.controller.js
@@ -1,41 +1,57 @@
 const DishModel = require("../models/dish.model");
 
 const createDish = async (req, res) => {
-  const { dishName, items, calories } = req.body;
-  if (!dishName || !items) {
-    return res.status(400).json({ message: "Please enter all fields" });
+  try {
+    const { dishName, items, calories } = req.body;
+    if (!dishName || !items) {
+      return res.status(400).json({ message: "Please enter all fields" });
+    }
+    const dishdata = await DishModel.create({ dishName, calories, items });
+    res.status(201).json({ message: "Dish created successfully", dishdata });
+  } catch (error) {
+    res.status(500).json({ message: error.message });
+
   }
-  const dishdata = await DishModel.create({ dishName, calories, items });
-  res.status(201).json({ message: "Dish created successfully", dishdata });
 };
 
 const getDish = async (req, res) => {
-  const dishdata = await DishModel.find();
-  res.status(200).json({ dishdata });
+  try {
+    const dishdata = await DishModel.find();
+    res.status(200).json({ dishdata });
+  } catch (error) {
+    res.status(500).json({ message: error.message });
+  }
 };
 
 const updateDish = async (req, res) => {
-  const id = req.params.id;
+  try {
+    const id = req.params.id;
+    const payload = req.body;
+    if (!id || !payload) {
+      return res.status(400).json({ message: "Please enter all fields" });
+    }
+    const data = await DishModel.findByIdAndUpdate(id, payload, { new: true });
 
-  const data = await DishModel.findByIdAndUpdate(id, req.body, { new: true });
-
-  res.status(200).json({
-    message: "Data Updated Sucessfully",
-    data
-  });
+    res.status(200).json({
+      message: "Data Updated Sucessfully",
+      data
+    });
+  } catch (error) {
+    res.status(500).json({ message: error.message });
+  }
 };
 
 const deleteDish = async (req, res) => {
-    
-    const id = req.params.id 
-
+  try {
+    const id = req.params.id
+    if (!id) return res.status(400).json({ message: "Please enter all fields" });
     await DishModel.findByIdAndDelete(id)
-
-    res.status(200).json({
-        message : "Data Deleted Succesfully"
-    })
+    res.status(200).json({message: "Data Deleted Succesfully"})
+  } catch (error) {
+    res.status(500).json({ message: error.message });
+  }
 
 
 }
 
-module.exports = { createDish, getDish , updateDish , deleteDish };
+module.exports = { createDish, getDish, updateDish, deleteDish };
